Ajusta tipo e migração do peso

+++ cadastro/models.py
@@ -13,7 +13,7 @@ class Animal(models.Model):
     ear_tag_number = models.CharField('nº do brinco', max_length=50)
     mother_ear_tag_number = models.CharField('nº do brinco da mãe', max_length=50)
     birth_date = models.DateField('data de nascimento', blank=True, null=True)
-    weight = models.DecimalField('peso (kg)', max_digits=6, decimal_places=2, blank=True, null=True)
+    weight = models.FloatField('peso (kg)', blank=True, null=True)
     weigh_date = models.DateField('data de pesagem', blank=True, null=True)
 
     def __str__(self):
